Implement required grammar and InlineTransformer
The required Lark grammar rules and the methods of the InlineTransformer
required by the instructions were implemented. The regular expressions
were reused from the last project (lispy-lex).

Independent tests were added to `parser.py`. They will not interfere
eith the official tests suite as they will be only executed when the
script is run on standalone mode.

The tests are now successfully executed.